chore: move `registerMempoolPromStats` fn to db helper module
